extract route as a file
